+++ requirements.txt
@@ -1,7 +1,7 @@
 asyncio==3.4.3
 discord.py==1.0.0
 Pillow==5.4.1
-psycopg2==2.7.7
+psycopg2==2.8.1
 PyNaCl==1.3.0
 request==2018.11.20
 requests==2.21.0
