+++ src/CharacterUtils.py
@@ -17,7 +17,7 @@
 ##    You should have received a copy of the GNU General Public License
 ##    along with this program. If not, see <http://www.gnu.org/licenses/>
 
-from DatabaseManager import *
+from src.DatabaseManager import *
 #from BotTools import DBJDR
 
 class Item:
